requirements.txt
numpy==1.22.1
Pillow==9.0.0
pytest==6.2.5
flake8==4.0.1